Refactor office to office_access
