+++ lib/ui/pages/home/office_access/access.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:fuks_app/generated/doorman/doorman.pb.dart';
 import 'package:fuks_app/services/doorman.dart';
-import 'package:fuks_app/ui/pages/home/office/access_time_snack_bar.dart';
+import 'package:fuks_app/ui/pages/home/office_access/access_time_snack_bar.dart';
 import 'package:fuks_app/ui/widgets/error.dart';
 import 'package:fuks_app/ui/widgets/constrained_list_view.dart';
 import 'package:undraw/undraw.dart';
